refactor: Follow atomic design gidelines
