+++ src/index.ts
@@ -1,1 +1,3 @@
-export * from './Button'
+export * from './atoms/Button'
+export * from './atoms/Select'
+export * from './organisms/ProductsSearchFilter'
